Deduplicate getIndexesForPlan and findCoveringIndexes (microsoft#51)
